Error when all data is not consumed
It is perfectly legal for a DFDL schema to not consume all the data when
parsing and still be considered valid. However, this is almost always
considered an error to users of a schema. So in this case, log that
there was remaining data and throw an exception, causing the FlowFile to
be transferred to the REL_FAILURE relationship.

DFDL-1797