Do not use the NiFi validator to ensure schema file existence
We currently use the FILE_EXISTS_VALIDATOR to ensure that the DFDL
Schema File property is set to a file that actually exists. However,
validation errors are not written to logs--they only visible in the NiFi
GUI and are are non-existent in MiNiFi. This can make it difficult to
determine what is wrong if there is a typo in the DFDL Schema File
property.

Instead, this just requires that the DFDL schema property is non-empty
and we try to compile the property value regardless. If the path does
not exist then we follow the normal schema compilation failure path,
which generates an error log message and the flow file routes to the
failure relationship.

Note that this means DFDL schema file existence is not checked until a
flow file is sent to the processor, and a log message is generated for
every flow file with the compilation error diagnostic.