Configure validation when the DataProcessor is created an cached
Calling the DataProcessor withValidation() function can have overhead
and can lead to the loss of Daffodil internal cached state used to
improve performance and reduce large memory allocations, related to uses
parsing with schemas that use lengthKind="pattern". The loss of this
state can lead to excessive allocations and OutOfMemory exceptions.

This moves the configuration of validation to when the DataProcessor is
created and stored in the cache. This ensures all calls to parse can
shared this state and avoid heavy allocations.

Additionally refactors the CacheKey to make it more clear it is about
all the things needed to create a DataProcessor, which includes the
validation mode.