Change how the release job is determined
Currently if a matrix configuration changes, it might be necessary to
update the 'Release Info' step logic to match. This is not odious since
the matrix configurations and 'Release Info' step are far apart.

To fix this, this includes a new is_release_job configuration to
existing matrix configs that is set immediately below the existing
matrix configurations, and is used in the 'Release Info' step. This
means that if matrix configurations change, it is more likely that we'll
notice to also update the is_release_job config if need.