Add support for setting DFDL external variables
The name/value of NiFi dynamic properties are now treated as DFDL
external variables. Property values are allowed to be NiFi expressions,
which are evaluated and set as the variable values. If a property value
evaluates to the empty string, it is ignored--this helps when a NiFi
expression determines that a variable does not apply for a schema, since
DFDL requires that all variables passed in externally must be valid for
the schema. This does mean it is not possible to have a variable with
the value of an empty string, but this should be rare and worked around.

This also refactored exceptions, so that functions throw correct
exceptions instead of turning everything into an IOException, which
worked but is not technically correct.

Also added a new "External Variables" section to additional details
pages documening this capability, and cleaned up the additional details
pages.