Support NiFi 2.x
The latest versions of NiFi change/remove API functions that are used by
the Daffodil processors. Trying to use the current version of these
processors in newer versions of Daffodil leads to NoClassDefFoundErrors
and breaks NiFi.

To fix this, this replaces those API functions/classes with variants
that exist in both older and newer versions of NiFi. This allows this
processor to maintain compatibility with older versions of NiFi while
still working with newer versions.

Specific changes include:

- NiFi 2.x removes support for EventDriven processors. This is likely
  rarely used in NiFi 1.x flows so support for EventDriven is removed
  for DaffodilParse/Unparse processors
- NiFi 2.x replaces the allowableValues(AllowableValues...) function
  with allowableValues(DescribedValue...). The DescribedValue variant
  does not exist in older versions of NiFi, so we instead just use the
  allowableValues(String...) variant, which exists in all versions of
  NiFi. This loses the human readable description of the field, but the
  allowable values are mostly self-descriptive. And in case there is
  still confusion, descriptions of the possible values are added to
  additionalDetails.html

Closes #18